tcpflow: update livecheck
The existing `livecheck` block for `tcpflow` now redirects to an AWS
S3 file listing page, which doesn't contain any tarball links in the
HTML as the file metadata is fetched client-side. This updates the
`livecheck` block to check the XML file containing the file
information and loosens the regex to be able to continue matching in
this new context.