git-annex: update build flags to enable p2http
This commit introduces two changes in the cabal build command, one is
a chore and and one enables additional functionality by adding an
optional build flag.

1. remove build dependency constraint

Following the TODO in the code comment, the pinned constraint is
removed. The underlying issue (yesodweb/yesod/issues/1854) is marked
as resolved and issue comments confirm that the bound on wai-extra is
no longer needed.

2. add Servant build flag enabling p2http

From git-annex.cabal:

> Use the servant library, enabling using annex+http urls and
> git-annex p2phttp

This flag and related features were introduced in git-annex
v10.20240731. In practice, git-annex built with this flag is able to
push annexed files over http(s). An example context where this is
useful are Git forges with annex support, such as
https://codeberg.org/forgejo-aneksajo, a soft fork of
https://forgejo.org.

This introduces additional build dependencies (servant-*, warp-*),
managed by cabal. When tested in the homebrew/ubuntu22.04:latest
container, this increased the binary size from 123M to 129M.

See:
- https://git-annex.branchable.com/git-annex-p2phttp/
- https://git-annex.branchable.com/tips/smart_http_server/