bash: depend on readline
This will fix issues like #229796, since our version of `readline` is
updated faster than the version of `readline` that is bundled with
`bash`.

The `readline` dependency was previously removed in #26260 because it's
nicer for a login shell to have no dependencies, but we already depend
on `ncurses` so that motivation doesn't really apply anymore.