gwyddion: update homepage, livecheck
The existing gwyddion.net URLs support HTTPS, so this updates the
homepage and `livecheck` block URLs accordingly. Besides that, this
also updates the `livecheck` block regex to achieve the same goal in
a way that's potentially less fragile (i.e., using `\s` instead of a
literal space and also allowing whitespace or the start of a trailing
HTML tag as the ending anchor).