damask-grid: add livecheck
By default, livecheck is unable to identify versions for `damask-grid`
and the formula also isn't excluded from autobump, so it's currently
failing with an `unable to get versions` error. This adds a
`livecheck` block that uses the `GithubLatest` strategy for now, as
the first-party website hasn't been updated to list 3.0.2 as the
newest version (a month on).

With this setup, livecheck could potentially surface a new version
before the tarball is available on the first-party website and
autobump would fail. However, it would eventually work when the file
for the new version becomes available, so that would only be a problem
if there's a long gap between those events. The alternative is to
check the source code installation page on the website that contains
the tarball URL but that's currently out of date, so we could be
stuck on an older version for some indeterminate amount of time. Of
those options, the former would ensure we update to the newest version
whenever it's ready, so I'm using `GithubLatest` here until the
first-party website reliably references the latest version.