checkstyle: add livecheck
By default livecheck checks the Git tags for `checkstyle` and the
newest tag is `checkstyle-11.0.0` but there hasn't been a release in
the two weeks since the tag was created and the website also
references 10.26.1 as the latest version. The formula uses a GitHub
release asset in the `stable` URL, so this addresses the issue by
adding a `livecheck` block that uses the `GithubLatest` strategy.