tree-sitter: split CLI to a separate formula
The CLI build dependencies are pretty heavy, which is unpleasant if
you're forced to build `tree-sitter` from source when you're most likely
only looking for the library.

This also removes `libtree-sitter` and its dependents from the
dependency trees of `rust` and `node`, which should make the dependent
tests for those formula just a tiny bit faster.