ca-certificates: improve Linux post-install
Let's check that the expected system CA certificates path exists and is
readable early so that, if it doesn't, we don't print the spurious
message about doing

    brew install openssl
    brew postinstall ca-certificates

This also allows us to avoid the early return in
`load_certificates_from_file`, which can risk silently ignoring the
Mozilla CA bundle we wanted to install if something went wrong.

The caveats have been updated accordingly.

While we're here:
- call `rm` with `force: true` so we don't need to check for existence
- use `Pathname#read` more consistently (as it's what's used elsewhere
  in this formula)