saxon: add gizmo script
In the saxon jar, there is actually a second tool to interactively perform simple operations such as examining the content of a document, renaming or deletion of selected elements in the document, or performing simple XSLT transformation and XSD validation.

This adds a shell script so that it is directly executable.

See here: https://www.saxonica.com/html/documentation10/gizmo/index.html