envoy: enable `wamr` and drop old workarounds
Previously, WASM was disabled as the default of V8 would often fail to
build on version bumps and across multiple runners due to compiler
versions. It looks like nixpkgs switched the default to WAMR due to
similar issues[^1] so trying this out in Homebrew formula too.

Also drop some old workarounds based on versions of projects used and
add a comment for when to drop `-Wno-deprecated-literal-operator`.

[^1]: https://github.com/NixOS/nixpkgs/commit/defb2298de66f6b32ec228a532dc6c8e34733b25