pdnsrec 5.3.0
`pdnsrec` is autobumped but the workflow is failing to update to
5.3.0 because it uses a tar.xz file instead of tar.bz2. This manually
updates the version/URL accordingly.

The Python dependency is required to resolve a build error: `Settings
table table.py has changed, python is needed to regenerate the related
settings files but python was not found. Please install python and
re-run configure`.