metabase 0.56.3
This updates `metabase` to the latest version, 0.56.3.

The existing `livecheck` block for `metabase` is returning an `Unable
to get versions` error, as the first-party download page for the OSS
jar only contains an unversioned link now (/latest/metabase.jar). This
updates the `livecheck` block to use the `GithubLatest` strategy, as
the GitHub release body text contains a versioned link to the OSS jar
file.