php: update livecheck
The existing `livecheck` block for `php` checks the download page but
it now gives an `Unable to get versions` error, as the page no longer
contains tarball links by default. This updates the URL to use the
`source=Y` query string parameter, which causes the download page to
include the source tarball links.