quasi88 0.7.3
This updates `quasi88` to the latest version, 0.7.3.

The existing `livecheck` block for `quasi88` checks the upstream
download page but it returns an `Unable to get versions` error, as
the source archive is now a zip file instead of a tarball. This
addresses the issue by updating the regex to also match a zip file.