gtk+3: update livecheck
The existing `livecheck` block for `gtk+3` uses a regex that expects
a file format like `gtk+-3.24.42.tar.xz`, where the prefix is
`gtk+-`. The current upstream files for `gtk+3` don't include a `+`
in the prefix (i.e., `gtk-`), so the check is failing with an `Unable
to get versions` error. This addresses the issue by updating the
`livecheck` block regex to make the `+` optional in the prefix.