manticoresearch: update livecheck
The existing `livecheck` block for `manticoresearch` uses the
`GithubLatest` strategy but it returns an `Unable to get versions`
error because the regex only matches versions with an even-numbered
patch (which 13.6.7 does not have). Everything I've seen upstream
suggests that 13.6.7 is the newest stable version, so removing this
restriction seems appropriate to me at this point.