otf2: update homepage, livecheck
The existing `livecheck` block for `otf2` is giving an `Unable to
get versions` error, as the homepage redirects to the overview page
and that doesn't contain an `otf2` tarball link. This updates the
`livecheck` block URL to check the download page, which is where the
`otf2` tarball link is found. Besides that, this updates the
`homepage` URL to resolve the redirection.