roxctl: update livecheck
The existing `livecheck` block for `roxctl` uses the `GithubLatest`
strategy but it's currently returning 4.6.9 as newest instead of
4.8.2. 4.6.9 was released after 4.8.2 and marked as "latest", so we
can't rely on the "latest" release always being the highest version
due to upstream creating releases for multiple major/minor versions
in an indeterminate order.

This addresses the issue by updating the `livecheck` block to use the
`GithubReleases` strategy, which finds the highest version from
recent releases. I checked for upstream pages that may link to the
latest versions in a more lightweight fashion but didn't find any
alternatives.