zsync: add livecheck
By default, livecheck uses the `Git` strategy for `zsync` but this
doesn't work because the repository doesn't have any tags. This adds
a `livecheck` block that checks the first-party download page, which
links to the stable tarball.

This was the `livecheck` block we were using before the formula was
deprecated (as zsync.moria.org.uk wasn't accessible at the time) and
the `livecheck` block was removed. The formula was revived with
version 0.6.3 (as the website works again) but the `livecheck` block
wasn't restored until now.