lsr: remove using argument
I added support to `DownloadStrategyDetector` for tangled.sh Git URLs,
so the `using: :git` argument in the `lsr` formula is (or will be)
redundant. This formula will fail `brew audit` (after we remove the
exception to skip the related audit for `lsr`), so this removes the
argument.

Besides that, this modifies the formula to use `zig@14`, as the build
will fail with errors otherwise (i.e., `expected function or variable
declaration after pub` for `build.zig` lines like
`pub usingnamespace @import("src/zeit.zig");` and
`pub usingnamespace zzdoc;`).