ola: configure `olad` service
Configures the `olad` service that comes with the `ola` formula.

Logging is by default done to `stderr`, thus only `error_log_path` is
needed. The default log level is Warnings.

The `olad` comes with an HTTP interface. It has a feature for exiting
the service by means of making a GET request to the `/quit` path. This
results in the `stopped` state for brew, requiring manual intervention
using `brew services stop ola` before it can be (re)started. The feature
is disabled using the command line argument `--no-http-quit`.