plakar: fix agent service
With v1.0.4, it is no longer necessary to start the agent manually with
`plakar agent`. The agent starts automatically and is killed after a few
seconds.

Also, for some reason, tests were disabled because the agent couldn't
start in the CI. Instead of disabling the test, plakar supports the
agent `-no-agent`. With `-no-agent`, plakar doesn't allow concurrent
requests which is not an issue for the CI.