yaml-language-server: downgrade to 1.18.0
`yaml-language-server` was upgraded to 1.19.0, presumably because
livecheck's `Npm` strategy reported 1.19.0 as newest. However, 1.19.0
is currently set as the "next" version on npm, whereas 1.18.0 is the
"latest" version. This issue will be addressed by changes to the `Npm`
strategy's behavior, so this downgrades the formula to 1.18.0 to align
with the "latest" npm version. I assume that 1.19.0 will eventually
become "latest" but this will save users from installing 1.19.0 in
the interim time.